+++ src/main/java/com/hmh/mmp/entity/CreditEntity.java
@@ -44,7 +44,7 @@ public class CreditEntity {
     private String creditName; // 내역
     private String creditPhotoName; // 사진명
     private String creditMemo;
-    private Long minusAsset; // 지출
+    private Double minusAsset; // 지출
     private Double rate; // 할부이자.
     private Integer month; // 할부
     private Double creditGet; // 캐쉬백
